create separate app.js and server.js
