+++ package.json
@@ -2,7 +2,7 @@
   "name": "employereviewapp",
   "version": "1.0.0",
   "description": "Task by ChainTech",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "jest",
     "dev": "nodemon index.js"
